lawyerfinder/apdjango/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.5
django-cors-headers==3.7.0
django-sslserver==0.22
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
fpdf==1.7.2
idna==2.9
isodate==0.6.0
lxml==4.6.2
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
reportlab==3.5.59
requests==2.23.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zeep==4.0.0
